update do main e correção de erros

+++ src/VideoTape.cpp
@@ -1,7 +1,7 @@
 #include <string>
 #include "../include/VideoTape.hpp"
 
-int VideoTape::getPrice() {
+float VideoTape::getPrice() {
     return _price;
 }
